+++ theme/src/shared/frontmatter/index.ts
@@ -1,0 +1,4 @@
+export * from './page.js'
+export * from './home.js'
+export * from './post.js'
+export * from './friends.js'
