+++ theme/src/shared/index.ts
@@ -1,5 +1,6 @@
 export * from './base.js'
-export * from './frontmatter.js'
+export * from './frontmatter/index.js'
 export * from './options/index.js'
-export * from './page.js'
+export * from './page-data.js'
 export * from './blog.js'
+export * from './navbar.js'
