+++ theme/src/shared/options/index.ts
@@ -35,5 +35,4 @@ export type PlumeThemeData = ThemeData<PlumeThemeLocaleData>
 
 export * from './locale.js'
 export * from './plugins.js'
-export * from './navbar.js'
 export * from './encrypt.js'
